data_type column type fix

+++ project/models.py
@@ -63,7 +63,7 @@ class ComparisonItem(db.Model):
 class DataType(db.Model):
     __tablename__ = 'data_type'
 
-    id = db.Column(Integer, primary_key=True)
+    id = db.Column(SmallInteger, primary_key=True)
     sort_type = db.Column(String(255), nullable=False)
     type_name = db.Column(String(255), nullable=False)
